refactor: move mock_crust and mock_parsec into mock module
